Adjust cart admin

+++ cart/middleware.py
@@ -1,3 +1,4 @@
+# cart/middleware.py
 from .models import Cart, CartItem
 
 class TransferCartMiddleware:
@@ -13,20 +14,21 @@ class TransferCartMiddleware:
 
             if old_session_key:
                 try:
-                    anonymous_cart = Cart.objects.get(session_key=old_session_key)
-                    user_cart, created = Cart.objects.get_or_create(user=request.user)
+                    anonymous_cart = Cart.objects.filter(session_key=old_session_key).first()
+                    if anonymous_cart:
+                        user_cart, _ = Cart.objects.get_or_create(user=request.user)
 
-                    for item in anonymous_cart.items.all():
-                        cart_item, created = CartItem.objects.get_or_create(cart=user_cart, product=item.product)
-                        if created:
-                            cart_item.quantity = item.quantity
-                        else:
-                            cart_item.quantity += item.quantity
-                        cart_item.price = item.price
-                        cart_item.promotion = item.promotion
-                        cart_item.save()
+                        for item in anonymous_cart.items.all():
+                            cart_item, created = CartItem.objects.get_or_create(cart=user_cart, product=item.product)
+                            if created:
+                                cart_item.quantity = item.quantity
+                            else:
+                                cart_item.quantity += item.quantity
+                            cart_item.price = item.price
+                            cart_item.promotion = item.promotion
+                            cart_item.save()
 
-                    anonymous_cart.delete()
+                        anonymous_cart.delete()
                 except Cart.DoesNotExist:
                     pass
